When linking files with multiple links, delete the one with fewer.
This helps avoid severing multiple links shared by files that
are outside the trees being processed.

For example, if you had identical files a, b and c and went
"rat a b; rat b c", it would first link a and b together, then
link b and c together, leaving a with a single link.